Rework tornado callbacks to ensure correct behavior
Previously we were just assuming that the currently executing callback
was the first one scheduled. If this isn't the case, then we may end
up missing messages.

Rework the way redis messages are dispatched so that all messages are
handled. This commit also halves the amount of tornado callbacks that
are scheduled for each redis message.